Creating new log middleware for the database connection and it's logs, and deleting cypress testing inside of frontend
