+++ backend/src/middlewares/loggerMongoMiddleware.js
@@ -1,0 +1,16 @@
+const winston = require('winston');
+
+// Logger configuration
+const logger = winston.createLogger({
+    level: 'info',
+    format: winston.format.combine(
+        winston.format.timestamp(),
+        winston.format.json()
+    ),
+    transports: [
+        new winston.transports.Console(),
+        new winston.transports.File({ filename: 'logs/db.log' })
+    ],
+});
+
+module.exports = logger;
