--- frontend/cypress/fixtures/customers.json
@@ -1,15 +1,0 @@
-[
-    {
-      "id": 1,
-      "firstname_owner": "Fernando",
-      "lastname_owner": "Martínez",
-      "phone_number": "1234567890"
-    },
-    {
-      "id": 2,
-      "firstname_owner": "José",
-      "lastname_owner": "Medina",
-      "phone_number": "0987654321"
-    }
-]
-
